% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_medline.R
\name{add_medline}
\alias{add_medline}
\title{Adding additional Medline info}
\usage{
add_medline(data_table, output_filename_base, type)
}
\arguments{
\item{data_table}{A data.frame that will supplement with additional Medline information.}

\item{output_filename_base}{A character same as filename_base from \code{\link{scrap_pubmed}}}

\item{type}{Either cited_by either references to specify data.frame.}
}
\value{
A data.frame with merged additional Medline info
}
\description{
Supplements the \code{data_table} with additional information from the Medline database.
}
