% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_PMC.R
\name{collapse_df}
\alias{collapse_df}
\title{Collapse uncollapsed dfs from \code{full_text} or \code{pic_captures}}
\usage{
collapse_df(dfs_list)
}
\arguments{
\item{dfs_list}{A character vector of PMCs.}
}
\value{
List of pictures captures of \code{PMC_list}.
}
\description{
Collapse uncollapsed dfs from \code{full_text} or \code{pic_captures}
}
