% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\docType{data}
\name{dictionary}
\alias{dictionary}
\title{20k most common words in the English language}
\format{An object of class \code{"data.frame"}}
\source{
\href{https://github.com/first20hours/google-10000-english}{google-10000-english}
}
\usage{
data(common_word_dictionary)
}
\description{
Data from Google's Trillion Word Corpus that contains a list of the 20,000
most common English words in order of frequency, as determined by n-gram frequency analysis.
}
\examples{
data(dictionary)
}
\references{
Modified version of the 20,000 most common English words in order of frequency,
as determined by n-gram frequency analysis of the Google's Trillion Word Corpus.
}
\keyword{datasets}
