% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_abstracts.R
\name{download_abstracts}
\alias{download_abstracts}
\title{Downloading abstracts by PMID list}
\usage{
download_abstracts(PMID_list)
}
\arguments{
\item{PMID_list}{A character vector PMIDs}
}
\value{
data.frame
}
\description{
Downloading abstracts by PMID list
}
