% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download "cited_in" or "references" data from PubMed}
\usage{
download_data(PMID, type = c("cited_by", "references"))
}
\arguments{
\item{PMID}{A single character or integer of interests paper.}

\item{type}{One of "cited_by" or "references" to specify output info.}
}
\value{
If available data is not \code{NULL} then returns corresponding to
\code{type} character vector.
}
\description{
Download data about papers that cites input paper from PubMed by its PMID.
}
