% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_text}
\alias{filter_text}
\title{Remove words in dictionary from \code{"character"} string}
\usage{
filter_text(x, d)
}
\arguments{
\item{x}{\code{"character"} string}

\item{d}{dictionary of commonly used words in the English language (e.g., \code{data("common_word_dictionary")})}
}
\value{
a vector of \code{"character"} strings excluding words found in
common word dictionary
}
\description{
Remove words in dictionary from \code{"character"} string
}
\examples{
data("common_word_dictionary")
x <- "Hello, my name is Shan Sabri!   less common words are: cells, differentation, reprogramming"
x <- strip_punctuation(x)
dictionary <- as.vector(dictionary$V1[1:5000])
filter_text(x, d = dictionary)
# # [1] "shan"           "sabri"          "cells"          "differentation" "reprogramming"
}
