% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_PMC.R
\name{full_text}
\alias{full_text}
\title{Get full text of PMC papers}
\usage{
full_text(PMC_list, Collapse = T)
}
\arguments{
\item{PMC_list}{A character vector of PMCs.}

\item{Collapse}{Need to collapse sections into a single text?}
}
\value{
List of full texts of \code{PMC_list}.
}
\description{
Get full text of PMC papers
}
