% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_foo_citation.R
\name{graph_foo_citation}
\alias{graph_foo_citation}
\title{Citation_by main function}
\usage{
graph_foo_citation(PMID_list, output_filename_base, n_cites_threshold = 2)
}
\arguments{
\item{PMID_list}{A character vector of PMIDs.}

\item{output_filename_base}{A character with path to output folder plus filename base.}

\item{n_cites_threshold}{An integer. The minimum number of papers that a given article should refer to.}
}
\value{
A data.frame with info about papers that refer to a given \code{PMID_list}.
}
\description{
Creates citation_by graph and corresponding data.frame with
additional info about the number of citations of each paper.
}
