% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_foo_references.R
\name{graph_foo_references}
\alias{graph_foo_references}
\title{References main function}
\usage{
graph_foo_references(PMID_list, output_filename_base, n_refers_treshold = 2)
}
\arguments{
\item{PMID_list}{A character vector of PMIDs.}

\item{output_filename_base}{A character with path to output folder plus filename base.}

\item{n_refers_treshold}{An integer. The minimum number of papers that a given article should refer to.}
}
\value{
A data.frame with info about papers that refer to a given \code{PMID_list}.
}
\description{
Creates references graph and corresponding data.frame with
additional info about the number of citations of each paper.
}
