% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_foo_similar_articles.R
\name{graph_foo_similar}
\alias{graph_foo_similar}
\title{Citation_by main function}
\usage{
graph_foo_similar(
  PMID_list,
  output_filename_base,
  n_similar_threshold = 2,
  year_left,
  year_right = substr(Sys.Date(), 1, 4)
)
}
\arguments{
\item{PMID_list}{A character vector of PMIDs.}

\item{output_filename_base}{A character with path to output folder plus filename base.}

\item{n_cites_threshold}{An integer. The minimum number of papers that a given article should be similar to.}
}
\value{
A data.frame with info about papers that are similar to a given \code{PMID_list}.
}
\description{
Creates similar articles graph and corresponding data.frame with
additional info about the number of citations of each paper.
}
