% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_PMC.R
\name{pic_captures}
\alias{pic_captures}
\title{Get pictures captures PMC papers}
\usage{
pic_captures(PMC_list, Collapse = T)
}
\arguments{
\item{PMC_list}{A character vector of PMCs.}

\item{Collapse}{Need to collapse sections into a single text?}
}
\value{
List of pictures captures of \code{PMC_list}.
}
\description{
Get pictures captures PMC papers
}
