% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{references}
\alias{references}
\title{Cited_by function}
\usage{
references(PMID_list)
}
\arguments{
\item{PMID_list}{A character vector of PMIDs of interests papers.}
}
\value{
If available data is not \code{NULL}.
then returns data.frame with data about papers on which refers \code{PMID_list}.
}
\description{
Download data about papers on which refers given \code{PMID_list} from PubMed
by PMID character vector with \code{\link{download_data}} and returns data.frame.
}
