% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_standalone_digits}
\alias{remove_standalone_digits}
\title{Strip standalone digits from \code{"character"} string}
\usage{
remove_standalone_digits(x)
}
\arguments{
\item{x}{a vector of \code{character} string}
}
\value{
a vector of \code{"character"} strings excluding standalong numeric digits
}
\description{
Strip standalone digits from \code{"character"} string
}
\examples{
data("common_word_dictionary")
x <- "Hello, my name is Shan Sabri!   less common words are: cells, differentation, reprogramming, 1 42, 101"
x <- strip_punctuation(x)
dictionary <- as.vector(dictionary$V1[1:5000])
x <- filter_text(x, d = dictionary)
remove_standalone_digits(x)
# [1] "shan"           "sabri"          "cells"          "differentation" "reprogramming"
}
