% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap_pubmed.R
\name{scrap_pubmed}
\alias{scrap_pubmed}
\title{Main function}
\usage{
scrap_pubmed(
  PMID_list,
  filename_base,
  n_cites_threshold = 2,
  n_refers_treshold = 2
)
}
\arguments{
\item{PMID_list}{A character vector of PMIDs of interesting papers.}

\item{filename_base}{A character contains path to output and filename base}

\item{n_cites_threshold}{An integer. The minimum number of papers that a given article should refer to.}

\item{n_refers_treshold}{An integer. The minimum number of papers that a given article should refer to.}
}
\value{
List of two data.frame's. First -- about papers that cites given \code{PMID_list}, second -- about papers on which refer given \code{PMID_list}.
}
\description{
Main function
}
