% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similar_articles.R
\name{similar_articles}
\alias{similar_articles}
\title{Similar_articles function}
\usage{
similar_articles(PMID_list, year_left, year_right = substr(Sys.Date(), 1, 4))
}
\arguments{
\item{PMID_list}{A character vector of PMIDs of interests papers.}

\item{year_left}{Shows left time interval for output}

\item{year_right}{Shows right time interval for output. Default - current year.}
}
\value{
If available data is not \code{NULL}.
then returns data.frame with data about papers that are similar to \code{PMID_list}.
}
\description{
Download data about papers that are similar to given \code{PMID_list} from PubMed
by PMID character vector with \code{\link{download_similar}} and returns data.frame.
}
