% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strip_punctuation}
\alias{strip_punctuation}
\title{Strip punctuation from character string}
\usage{
strip_punctuation(x)
}
\arguments{
\item{x}{\code{"character"} string}
}
\value{
\code{"character"} string in lowercase without punctuation
}
\description{
Strip punctuation from character string
}
\examples{
x <- "Hello, my name is Shan Sabri!   "
strip_punctuation(x)
# "hello my name is shan sabri"
}
