% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_medline.R
\name{add_medline}
\alias{add_medline}
\title{Adding additional Medline info}
\usage{
add_medline(data_table, venv)
}
\arguments{
\item{data_table}{A data.frame that will supplement with additional Medline information.}

\item{venv}{A character specifies your python's venv in which installed biopython and pandas.}
}
\value{
A data.frame with merged additional Medline info
}
\description{
Supplements the \code{data_table} with additional information from the Medline database.
}
